refactor: examples/7Guis/<example>/ sub-folders + relocate example tooling
Two restructures bundled because they share the spec.cjs require-path
sweep and want to land atomically (each one in isolation would leave the
test runner broken).

7Guis cluster:
  - examples/seven_guis/<file>          ->  examples/7Guis/<example>/<file>
  - Each 7GUIs task is now its own self-contained sub-folder with its
    cljs source, hand-written index.html, and Playwright spec — mirroring
    counter/, todomvc/, login/, etc.
  - Directory name is `7Guis` (matching the original 7GUIs.com
    capitalisation). CLJS namespaces can't start with a digit, so the
    on-disk parent and the namespace tree diverge: each example takes
    its own top-level ns pair (cells.cells, crud.crud, ...) under a
    dedicated `examples/7Guis` shadow-cljs source root.

Example tooling moved out of implementation/scripts/:
  - implementation/scripts/run-examples-tests.cjs           ->  examples/scripts/
  - implementation/scripts/serve-and-run-examples-tests.cjs ->  examples/scripts/
  - implementation/scripts/spec-helpers.cjs                 ->  examples/scripts/
  - run-examples-tests.cjs resolves `playwright` explicitly
    out of implementation/node_modules so it works regardless of cwd.
  - The orchestrator continues to spawn shadow-cljs and http-server
    with cwd: IMPL_ROOT, so npm scripts are still invoked from the
    implementation tree. No examples/package.json — the implementation
    devDependencies cover both worlds.
  - implementation/package.json's `test:examples` script now points at
    ../examples/scripts/serve-and-run-examples-tests.cjs.

Sweep:
  - 12 spec.cjs files updated (require path:
    '../../implementation/scripts/spec-helpers.cjs' becomes
    '../scripts/spec-helpers.cjs' for top-level examples and
    '../../scripts/spec-helpers.cjs' for the 7Guis sub-folders).
  - implementation/shadow-cljs.edn: added '../examples/7Guis' source
    path; updated 6 :init-fn entries (seven-guis.<name>/run becomes
    <name>.<name>/run).

Verified locally: clojure -M:test (core, schemas), npm run test:cljs,
npm run test:browser, npm run test:elision, npm run test:examples
(15/15 PASS). bd: rf2-gqm7.
