+++ examples/7Guis/crud/crud.spec.cjs
@@ -16,7 +16,7 @@
  * size-6 multi-row <select> that backs the list.
  */
 
-const { expectVisible } = require('../../implementation/scripts/spec-helpers.cjs');
+const { expectVisible } = require('../../scripts/spec-helpers.cjs');
 
 module.exports = {
   name: 'crud (7guis #5)',
