+++ examples/scripts/spec-helpers.cjs
