+++ examples/state_machine_walkthrough/state_machine_walkthrough.spec.cjs
@@ -23,7 +23,7 @@ const {
   expectTextEquals,
   expectTextContains,
   expectVisible,
-} = require('../../implementation/scripts/spec-helpers.cjs');
+} = require('../scripts/spec-helpers.cjs');
 
 module.exports = {
   name: 'state-machines walkthrough (lockout)',
