+++ implementation/package.json
@@ -16,6 +16,6 @@
     "test:cljs": "shadow-cljs compile node-test && node out/node-test.js",
     "test:browser": "shadow-cljs compile browser-test && node scripts/serve-and-run-browser-tests.cjs",
     "test:elision": "shadow-cljs release elision-probe elision-probe-control && node scripts/check-elision.cjs",
-    "test:examples": "node scripts/serve-and-run-examples-tests.cjs"
+    "test:examples": "node ../examples/scripts/serve-and-run-examples-tests.cjs"
   }
 }
